+++ src/main/java/com/example/demo/repository/TratamientoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entity.Tratamiento;
 
-public interface TratamientoRepository extends CrudRepository<Tratamiento, Integer> {
+public interface TratamientoRepository extends CrudRepository<Tratamiento, Long> {
 
 }
